Bump pcapOffset from int32 to int64
This commit bumps the pcapOffset var from int32 to int64 to match changes in protobuf/steno.proto (see previous commit for details).

+++ rpc/rpc.go
@@ -83,7 +83,7 @@ func (s *stenographerServer) RetrievePcap(
                 return nil
         }
 
-        var pcapOffset int32 = 0
+        var pcapOffset int64 = 0
         buffer := make([]byte, chunkSize)
         for pcapOffset < maxSize {
                 if pcapOffset >= s.rpcCfg.PcapLimitSize {
